{"schema":"decodex/commit/1","summary":"Consolidate cohesive Swift bridge modules","authority":"manual"}
